All tests except for naive bounded backwards now pass. All cloud search functions now leave matrix unmodified.

+++ src/objects/clock.c
@@ -81,10 +81,10 @@ double CLOCK_Stop(CLOCK*cl)
 }
 
 /* Convert duration from ticks to msec */
-float CLOCK_Secs(CLOCK*cl)
+double CLOCK_Secs(CLOCK*cl)
 {
    cl->duration = cl->stop - cl->start;
-   return (float)(cl->duration);
+   return (double)(cl->duration);
 }
 
 /* dependencies for Get_RealTime() */
